pyproject.toml
[tool.poetry]
name = "tpe-redes"
version = "0.1.0"
description = ""
authors = ["Juan Negro <jnegro@itba.edu.ar>", "Mateo Bartellini Huapalla <mbartellini@itba.edu.ar>", "Juan Martín Barmasch <jbarmasch@itba.edu.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
